state: Basic buffer implementation with public API implementation
Inner implementation not in buffer_api.zig is designed to be as
minimal as possible in order to provide just the very primitives.
This should allow to relatively easily implement alternative
implementations for a text buffer.

buffer_api.zig has all the public API functions that will be available
to the extension language or the client.